@@ -17,64 +17,64 @@
     "build:aot": "ngc -p ./tsconfig-aot.json"
   },
   "dependencies": {
-    "@angular/animations": "4.0.1",
-    "@angular/common": "4.0.1",
-    "@angular/compiler": "4.0.1",
-    "@angular/compiler-cli": "~4.0.1",
-    "@angular/core": "4.0.1",
-    "@angular/flex-layout": "~2.0.0-rc.1",
-    "@angular/forms": "4.0.1",
-    "@angular/http": "4.0.1",
-    "@angular/platform-browser": "4.0.1",
-    "@angular/platform-browser-dynamic": "4.0.1",
-    "@angular/platform-server": "~4.0.1",
-    "@angular/router": "4.0.1",
-    "lodash": "^4.17.4",
-    "rxjs": "5.0.1",
-    "zone.js": "^0.8.9"
+    "@angular/animations": "~5.0.0",
+    "@angular/common": "~5.0.0",
+    "@angular/compiler": "~5.0.0",
+    "@angular/compiler-cli": "~5.0.0",
+    "@angular/core": "~5.0.0",
+    "@angular/forms": "~5.0.0",
+    "@angular/http": "~5.0.0",
+    "@angular/platform-browser": "~5.0.0",
+    "@angular/platform-browser-dynamic": "~5.0.0",
+    "@angular/platform-server": "~5.0.0",
+    "@angular/router": "~5.0.0",
+    "@types/node": "^8.0.47",
+    "lodash": "~4.17.4",
+    "rxjs": "~5.5.2",
+    "zone.js": "~0.8.18"
   },
   "devDependencies": {
-    "@ngtools/webpack": "^1.3.1",
-    "@types/jasmine": "^2.5.46",
-    "@types/lodash": "^4.14.64",
-    "angular2-router-loader": "^0.3.2",
-    "angular2-template-loader": "^0.6.0",
-    "awesome-typescript-loader": "^3.0.4",
-    "babel": "^6.5.2",
-    "babel-core": "^6.21.0",
-    "babel-loader": "^6.2.10",
-    "babel-preset-es2015": "^6.18.0",
-    "clean-webpack-plugin": "^0.1.16",
-    "css-loader": "^0.26.1",
-    "del": "^2.2.2",
-    "del-cli": "^0.2.0",
-    "es6-promise": "^4.0.5",
-    "es6-shim": "^0.35.2",
-    "extract-text-webpack-plugin": "^2.1.0",
-    "file-loader": "^0.9.0",
-    "html-loader": "^0.4.4",
-    "html-webpack-plugin": "^2.26.0",
-    "http-server": "^0.9.0",
-    "istanbul-instrumenter-loader": "^2.0.0",
-    "jasmine-core": "^2.5.2",
-    "karma": "1.6.0",
-    "karma-chrome-launcher": "^2.0.0",
-    "karma-coverage": "^1.1.1",
-    "karma-coverage-istanbul-reporter": "^1.0.0",
-    "karma-jasmine": "^1.1.0",
-    "karma-phantomjs-launcher": "^1.0.2",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^2.0.3",
-    "node-sass": "^4.4.0",
-    "null-loader": "^0.1.1",
-    "optimize-css-assets-webpack-plugin": "^1.3.1",
-    "raw-loader": "^0.5.1",
-    "sass-loader": "^4.1.1",
-    "style-loader": "^0.13.1",
-    "typescript": "2.3.2",
-    "webpack": "2.4.1",
-    "webpack-dev-server": "^2.4.5",
-    "webpack-merge": "^4.1.0",
-    "webpack-stream": "^3.2.0"
+    "@ngtools/webpack": "~1.8.0",
+    "@types/jasmine": "~2.6.2",
+    "@types/lodash": "~4.14.82",
+    "angular2-router-loader": "~0.3.5",
+    "angular2-template-loader": "~0.6.2",
+    "awesome-typescript-loader": "~3.3.0",
+    "babel": "~6.23.0",
+    "babel-core": "~6.26.0",
+    "babel-loader": "~7.1.2",
+    "babel-preset-es2015": "~6.24.1",
+    "clean-webpack-plugin": "~0.1.17",
+    "css-loader": "~0.28.7",
+    "del": "~3.0.0",
+    "del-cli": "~1.1.0",
+    "es6-promise": "~4.1.1",
+    "es6-shim": "~0.35.3",
+    "extract-text-webpack-plugin": "~3.0.2",
+    "file-loader": "~1.1.5",
+    "html-loader": "~0.5.1",
+    "html-webpack-plugin": "~2.30.1",
+    "http-server": "~0.10.0",
+    "istanbul-instrumenter-loader": "~3.0.0",
+    "jasmine-core": "~2.8.0",
+    "karma": "~1.7.1",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage": "~1.1.1",
+    "karma-coverage-istanbul-reporter": "~1.3.0",
+    "karma-jasmine": "~1.1.0",
+    "karma-phantomjs-launcher": "~1.0.4",
+    "karma-sourcemap-loader": "~0.3.7",
+    "karma-webpack": "~2.0.5",
+    "node-sass": "~4.6.0",
+    "null-loader": "~0.1.1",
+    "optimize-css-assets-webpack-plugin": "~3.2.0",
+    "raw-loader": "~0.5.1",
+    "sass-loader": "~6.0.6",
+    "style-loader": "~0.19.0",
+    "typescript": "~2.6.1",
+    "webpack": "~3.8.1",
+    "webpack-dev-server": "~2.9.4",
+    "webpack-merge": "~4.1.1",
+    "webpack-stream": "~4.0.0"
   }
 }
