@@ -1,6 +1,6 @@
 import { ReactiveFormsModule } from '@angular/forms';
 import { BrowserModule } from '@angular/platform-browser';
-import { HttpModule } from '@angular/http';
+import {HttpClientModule} from '@angular/common/http';
 import { NgModule } from '@angular/core';
 
 import { AppComponent } from './app.component';
@@ -21,7 +21,7 @@ import { SharedModule } from './shared/shared.module';
     routing,
     SharedModule,
     ReactiveFormsModule,
-    HttpModule,
+    HttpClientModule,
     AccountModule,
     AuthModule,
     WraithModule
