@@ -1,8 +1,8 @@
-import { Component, OnInit } from '@angular/core';
-import { ButtonListItem } from '../../components/button-list/button-list-item/button-list-item';
-import { Router } from '@angular/router';
-import { AccountService } from '../account-service/account.service';
-import { Account } from '../account.model';
+import {Component, OnInit} from '@angular/core';
+import {ButtonListItem} from '../../components/button-list/button-list-item/button-list-item';
+import {Router} from '@angular/router';
+import {AccountService} from '../account-service/account.service';
+// import {Account} from '../account.model';
 
 @Component({
   selector: 'cash-account-list',
@@ -17,18 +17,14 @@ export class AccountListComponent implements OnInit {
   }
 
   ngOnInit(): void {
-    let accounts: Account[] = [];
-    this.accountService.getAccounts().subscribe(
-      (result) => {
-        accounts = result;
-        accounts.forEach((account) => {
-          let buttonItem = new ButtonListItem(account);
-          buttonItem.addLabel(account.name);
-          buttonItem.addLabel(account.balance.toString(), account.balance >= 0 ? 'forestgreen' : 'indianred');
-          this.buttons.push(buttonItem);
-        })
-      }
-    );
+    this.accountService.getAccounts().subscribe((accounts: any) => {
+      accounts.forEach((account:any) => {
+        let buttonItem = new ButtonListItem(account);
+        buttonItem.addLabel(account.name);
+        buttonItem.addLabel(account.balance.toString(), account.balance >= 0 ? 'forestgreen' : 'indianred');
+        this.buttons.push(buttonItem);
+      });
+    });
   }
 
   onAccountAdd(): void {
