@@ -1,54 +1,41 @@
-import { Headers, Http, Response } from '@angular/http';
 import { Injectable } from '@angular/core';
 import { User } from '../user.model';
 import { Observable } from 'rxjs';
-import 'rxjs/Rx';
 import { BaseRequestService } from '../../shared/base-request.service';
+import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { FlashService } from '../../components/flash/flash.service';
 
 @Injectable()
 export class AuthService extends BaseRequestService {
 
-  constructor(private http: Http, private flashService: FlashService) {
+  constructor(private httpClient: HttpClient, private flashService: FlashService) {
     super();
   }
 
-  signup = (user: User) => {
+  signup(user: User): Observable<any> {
     const body = JSON.stringify(user);
-    const headers = new Headers({
-      'Content-Type': 'application/json'
-    });
-    return this.http.post(this.baseUrl + '/auth/signup', body, { headers: headers })
-      .map((response: Response) => response.json())
-      .map(data => {
-        if (data) {
-          localStorage.setItem('token', data.result[0].token);
-          return true;
-        }
-      })
-      .catch((error: Response) => {
-        return Observable.throw(error.json())
-      })
+    return this.httpClient.post(this.baseUrl + '/auth/signup', body, { headers: super.getHeader() })
+      .map((data: any) => {
+        localStorage.setItem('token', data.result[0].token);
+        return true;
+      }).catch((error: HttpErrorResponse) => {
+        this.flashService.error(error.name + ' (' + error.message + ')');
+        return Observable.throw(error)
+      });
   };
 
   signin = (user: User) => {
     const body = JSON.stringify(user);
-    const headers = new Headers({
-      'Content-Type': 'application/json'
-    });
-    return this.http.post(this.baseUrl + '/auth/signin', body, { headers: headers })
-      .map((response: Response) => response.json())
-      .map(data => {
-        if (data) {
+    return this.httpClient.post(this.baseUrl + '/auth/signin', body, { headers: super.getHeader() })
+      .subscribe((data: any) => {
           localStorage.setItem('token', data.result[0].token);
           return true;
-        }
-      })
-      .catch((error: Response) => {
-        localStorage.clear();
-        this.flashService.error(error.json().result[0].error.message);
-        return Observable.throw(error.json())
-      });
+        },
+        (error) => {
+          localStorage.clear();
+          this.flashService.error(error.message);
+          return false
+        });
   };
 
   logout = () => {
@@ -60,7 +47,13 @@ export class AuthService extends BaseRequestService {
   };
 
   getFullName = () => {
-    let user = JSON.parse(window.atob(localStorage.getItem('token').split('.')[1])).user;
-    return (user.firstName + ' ' + user.lastName);
+    let token = localStorage.getItem('token');
+    //If a token exists, the proceed to parse it.
+    if (token) {
+      let user = JSON.parse(window.atob(token.split('.')[1])).user;
+      return (user.firstName + ' ' + user.lastName);
+    } else {
+      return '';
+    }
   }
 }
